+++ src/main/java/com/arom/polisee/global/login/config/KakaoConfig.java
@@ -1,0 +1,18 @@
+package com.arom.polisee.global.login.config;
+
+import lombok.Getter;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+@Component
+@Getter
+public class KakaoConfig {
+
+    @Value("${spring.kakao.client-id}")
+    private String clientId;
+
+    @Value("${spring.kakao.redirect-uri}")
+    private String redirectUri;
+
+}
