+++ src/main/java/com/arom/polisee/global/login/entity/User.java
@@ -1,0 +1,40 @@
+package com.arom.polisee.global.login.entity;
+
+import jakarta.persistence.*;
+import lombok.*;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+
+import java.time.LocalDateTime;
+
+@Getter
+@Setter
+@Entity
+@NoArgsConstructor
+@AllArgsConstructor
+@Table(name = "users")  //  테이블 이름 명확히 지정
+@Builder
+@ToString
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)  // AUTO_INCREMENT & PRIMARY KEY
+    @Column(name = "user_id")
+    private Long id;
+
+    @Column(name = "kakao_id", nullable = false, unique = true)
+    private Long kakaoId;
+
+    @Column(name = "user_name", nullable = false)
+    private String userName;
+
+    @Column(name = "created_at", updatable = false)
+    @CreationTimestamp
+    private LocalDateTime createdAt;
+
+    @Column(name = "updated_at")
+    @UpdateTimestamp
+    private LocalDateTime updatedAt;
+
+
+}
