+++ src/main/java/com/arom/polisee/global/login/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package com.arom.polisee.global.login.repository;
+
+import com.arom.polisee.global.login.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByKakaoId(Long kakaoId);
+}
