--- src/main/java/com/arom_policy/demo/login/common/BaseException.java
@@ -1,14 +1,0 @@
-package com.arom_policy.demo.login.common;
-
-public class BaseException extends RuntimeException {
-    private final BaseResponseStatus status;
-
-    public BaseException(BaseResponseStatus status) {
-        super(status.getMessage()); // 에러 메시지를 가져옴
-        this.status = status;
-    }
-
-    public BaseResponseStatus getStatus() {
-        return status;
-    }
-}
