--- src/main/java/com/arom_policy/demo/login/config/KakaoConfig.java
@@ -1,18 +1,0 @@
-package com.arom_policy.demo.login.config;
-
-import lombok.Getter;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-@Component
-@Getter
-public class KakaoConfig {
-
-    @Value("${spring.kakao.client-id}")
-    private String clientId;
-
-    @Value("${spring.kakao.redirect-uri}")
-    private String redirectUri;
-
-}
