--- src/main/java/com/arom_policy/demo/login/dto/LoginResponseDto.java
@@ -1,13 +1,0 @@
-package com.arom_policy.demo.login.dto;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class LoginResponseDto {
-    private Long userId;
-    private String username;
-}
