--- src/main/java/com/arom_policy/demo/login/service/TokenService.java
@@ -1,88 +1,0 @@
-package com.arom_policy.demo.login.service;
-
-import com.arom_policy.demo.login.config.KakaoConfig;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.*;
-import org.springframework.stereotype.Service;
-import org.springframework.util.LinkedMultiValueMap;
-import org.springframework.util.MultiValueMap;
-import org.springframework.web.bind.annotation.RequestHeader;
-import org.springframework.web.client.RestTemplate;
-
-import java.util.HashMap;
-
-@Service
-@RequiredArgsConstructor
-public class TokenService {
-    private final KakaoConfig kakaoConfig;
-    private final RestTemplate restTemplate = new RestTemplate();
-    private final ObjectMapper objectMapper = new ObjectMapper();
-
-    public String getKakaoAccessToken(String code) {
-        String requestURL = "https://kauth.kakao.com/oauth/token";
-
-        MultiValueMap<String, String> params = new LinkedMultiValueMap<>();
-        params.add("grant_type", "authorization_code");
-        params.add("client_id", kakaoConfig.getClientId());
-        params.add("redirect_uri", kakaoConfig.getRedirectUri());
-        params.add("code", code);
-
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(MediaType.APPLICATION_FORM_URLENCODED);
-
-        HttpEntity<MultiValueMap<String, String>> requestEntity = new HttpEntity<>(params, headers);
-
-        try {
-            ResponseEntity<String> response = restTemplate.exchange(
-                    requestURL, HttpMethod.POST, requestEntity, String.class
-            );
-
-            JsonNode root = objectMapper.readTree(response.getBody());
-            return root.path("access_token").asText();
-        } catch (Exception e) {
-            e.printStackTrace();
-            throw new RuntimeException("카카오 토큰 요청 실패", e);
-        }
-    }
-
-    public HashMap<String, Object> getUserInfoFromToken(@RequestHeader("Authorization") String accessToken) {
-        HashMap<String, Object> userInfo = new HashMap<>();
-        String url = "https://kapi.kakao.com/v2/user/me";
-
-        // HTTP 요청 헤더 설정
-        HttpHeaders headers = new HttpHeaders();
-        headers.set("Authorization", "Bearer " + accessToken);
-        headers.setContentType(MediaType.APPLICATION_FORM_URLENCODED);
-
-        HttpEntity<String> requestEntity = new HttpEntity<>(headers);
-
-        try {
-            // API 요청 실행
-            ResponseEntity<String> response = restTemplate.exchange(url, HttpMethod.POST, requestEntity, String.class);
-
-            // JSON 응답 파싱
-            JsonNode root = objectMapper.readTree(response.getBody());
-            String username = root.path("properties").path("nickname").asText();
-            Long kakaoId = root.path("id").asLong();
-
-            // 사용자 정보 저장
-            userInfo.put("username", username);
-            userInfo.put("id", kakaoId);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-
-        return userInfo;
-    }
-
-    public String getKakaoAuthUrl() {
-        return "https://kauth.kakao.com/oauth/authorize"
-                + "?client_id=" + kakaoConfig.getClientId()
-                + "&redirect_uri=" + kakaoConfig.getRedirectUri()
-                + "&response_type=code"
-                + "&scope=profile_nickname";
-
-    }
-}
